refactor: complete focused facade architecture implementation
## Summary
- Finalize the focused facade architecture by fixing remaining inconsistent imports
- Remove the deprecated utils/index.ts barrel file completely
- Ensure all imports use proper focused facades throughout the codebase

## Changes Made
### Import Consistency Fixes
- Fix build_device.ts to use utils/build/index.ts instead of direct build-utils.ts import
- Fix test-common.ts to use utils/build/index.ts for executeXcodeBuildCommand import
- Update test files to use utils/responses/index.ts for SystemError imports instead of barrel
- Update scaffold project files to use proper focused facades for dynamic imports

### Architecture Completion
- **REMOVED**: src/utils/index.ts - deprecated barrel file completely deleted
- All imports now consistently use focused facades (build/, responses/, execution/, etc.)
- No more barrel imports anywhere in the codebase

## Technical Benefits
- **Improved tree-shaking**: Bundlers can better eliminate unused code
- **Clearer dependencies**: Import paths explicitly show what functionality is needed
- **Prevention of circular dependencies**: Focused facades reduce coupling risks
- **Better performance**: Eliminates loading of unused modules during startup
- **Architectural consistency**: Single pattern enforced across entire codebase

## Validation
- ✅ TypeScript compilation: No errors
- ✅ ESLint: Only 1 minor unrelated warning
- ✅ Prettier: All files properly formatted
- ✅ Tests: 1046 tests passing
- ✅ Build: Successful compilation with all tools working

The focused facade architecture is now fully implemented and enforced.
