--- src/utils/index.ts
@@ -1,21 +1,0 @@
-/**
- * This barrel will be removed in a future release after a deprecation period.
- */
-export * from './logger.ts';
-export * from './command.ts';
-export * from './CommandExecutor.ts';
-export * from './FileSystemExecutor.ts';
-export * from './validation.ts';
-export * from './errors.ts';
-export * from './build-utils.ts';
-export * from './xcode.ts';
-export * from './axe-helpers.ts';
-export * from './log_capture.ts';
-export * from './template-manager.ts';
-export * from './test-common.ts';
-export * from './xcodemake.ts';
-export * from './environment.ts';
-export * from './tool-registry.ts';
-export * from '../version.ts';
-export * from '../core/dynamic-tools.ts';
-export * from '../core/plugin-registry.ts';
